fixed an ambiguous cast to QVariant
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1873 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/qt/ui/UIPadSetting.cpp
@@ -95,7 +95,7 @@ void UIPadSetting::setPadKey( u32 key )
 		.arg( PERPAD )
 		.arg( mPadKey );
 	
-	QtYabause::settings()->setValue( settingsKey, key );
+	QtYabause::settings()->setValue( settingsKey, (uint) key );
 	mPadButtons.key( mPadKey )->setIcon( QIcon( ":/actions/icons/actions/button_ok.png" ) );
 	mPadButtons.key( mPadKey )->setChecked( false );
 	lInfos->clear();
